+++ src/async-concat.mjs
@@ -1,2 +1,7 @@
-import asyncChain from './async-chain'
-export default asyncChain
+import { ensureAsyncIterable } from './internal/async-iterable'
+
+export default async function * asyncConcat (...arrayOfIter) {
+  for (const iterable of arrayOfIter) {
+    yield * ensureAsyncIterable(iterable)
+  }
+}