+++ src/async-zip-all.mjs
@@ -1,2 +1,17 @@
-import asyncZipLongest from './async-zip-longest'
-export default asyncZipLongest
+import { ensureAsyncIterable } from './internal/async-iterable'
+
+export default async function * asyncZipAll (...iterables) {
+  const iters = iterables.map(arg => ensureAsyncIterable(arg)[Symbol.asyncIterator]())
+  try {
+    while (true) {
+      const results = await Promise.all(iters.map((iter) => iter.next()))
+      const done = results.every((r) => r.done)
+      if (done) return
+      yield results.map((r) => r.done ? undefined : r.value)
+    }
+  } finally {
+    for (const iter in iters) {
+      if (typeof iter.return === 'function') await iter.return()
+    }
+  }
+}