+++ src/chain.mjs
@@ -1,7 +1,0 @@
-import { ensureIterable } from './internal/iterable'
-
-export default function * chain (...arrayOfIter) {
-  for (const iterable of arrayOfIter) {
-    yield * ensureIterable(iterable)
-  }
-}