+++ src/concat.mjs
@@ -1,2 +1,7 @@
-import chain from './chain'
-export default chain
+import { ensureIterable } from './internal/iterable'
+
+export default function * concat (...arrayOfIter) {
+  for (const iterable of arrayOfIter) {
+    yield * ensureIterable(iterable)
+  }
+}